requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyMuPDF==1.23.8
python-docx==0.8.11
googletrans==4.0.0rc1
requests==2.31.0
Werkzeug==2.3.7
beautifulsoup4==4.12.2
openai==1.3.7
python-dateutil==2.8.2
schedule==1.2.0
